Nice changes: - Final Report almost finished - Student info almos finished
André Creppe
26-07-19

+++ database/migrations/2014_14_12_000002_create_final_reports_table.php
@@ -33,7 +33,7 @@ class CreateFinalReportsTable extends Migration
             $table->integer('grade_4_a');
             $table->integer('grade_4_b');
             $table->integer('grade_4_c');
-            $table->integer('final_grade');
+            $table->float('final_grade');
 
             $table->integer('hours_completed');
             $table->date('end_date');
